initial commit
i need to add the full $PAGE_THIS BASENAME line. if you logged in, it
drops right through. if not the form shows up but i need to fix the
logic of the inner if loop. i think we have the rest though. then we can
drop this include in to any page header.